requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
DateTime==5.5
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-select2==8.2.0
djangorestframework==3.15.2
idna==3.7
kombu==5.3.7
pillow==10.4.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
random-string==1.0
requests==2.32.3
setuptools==71.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==6.4.post2
